Move VersionChanges directive to sphinx.domains.changeset
